Port junit framwork for scripts from repo
This ports the custom framework for running junit tests via script files
from the repository module to common. The tests for the logger module were
used as the benchmark and have already been migrated to this type of
testing, including extension of the test scope and minor fixes in the
framework for module loading.